Update main.py
If you gave main it's REPO_ID and REPO_NAME directly in main, and it's token as an env key then the prev solution failed. This does not matter for github because there the user does not have option of giving REPO_ID and REPO_NAME as arguments. Basically, we added REPO_NAME = sys.argv[2] etc to the public version and this crashed in a bad way.